fix(drv8833): replace PRIu32 with %lu cast to avoid inttypes.h dependency
Agent-Logs-Url: https://github.com/miguel-sergio/esp32-smart-cooking-firmware/sessions/3cbd8276-1a72-4a2e-9e98-246c6f062c2c

+++ components/drv8833/drv8833.c
@@ -148,7 +148,7 @@ esp_err_t drv8833_ramp_to_speed(drv8833_dev_t *dev, drv8833_channel_t ch,
     for (uint32_t step = 1; step <= steps; step++) {
         int8_t v = (int8_t)(start + diff * (int)step / (int)steps);
         ESP_RETURN_ON_ERROR(drv8833_set_speed(dev, ch, v),
-                            TAG, "ramp set_speed failed at step %"PRIu32, step);
+                            TAG, "ramp set_speed failed at step %lu", (unsigned long)step);
         if (step < steps) {
             vTaskDelay(pdMS_TO_TICKS(RAMP_STEP_MS));
         }
